refactor: Reservation에 관련된 클래스 패키지 계층화
